feat: enable button only after validation and checkmark checked
